requirements.txt
langchain
langchain-ollama
langchain-chroma
chromadb
pandas
flask
psycopg2-binary